requirements.txt
streamlit==1.27.2
tensorflow==2.13.0
Pillow==10.0.0
numpy
